+++ src/models/internals/build-context-static-options.ts
@@ -1,0 +1,16 @@
+import { LoggerBase, LogLevelString, } from '../../utils';
+
+import { AngularBuildConfigInternal } from './angular-build-config-internal';
+
+export interface BuildContextStaticOptions {
+    workspaceRoot: string;
+    filteredConfigNames?: string[];
+    startTime?: number;
+    angularBuildConfig?: AngularBuildConfigInternal;
+    fromBuiltInCli?: boolean;
+    cliRootPath?: string;
+    cliVersion?: string;
+    cliIsGlobal?: boolean;
+    logger?: LoggerBase;
+    logLevel?: LogLevelString;
+}
