+++ src/models/internals/build-options-internal.ts
@@ -1,0 +1,5 @@
+import { BuildOptions } from '../build-options';
+
+export interface BuildOptionsInternal extends BuildOptions {
+    environment: { [key: string]: boolean | string };
+}
