+++ src/models/internals/global-script-style-parsed-entry.ts
@@ -1,0 +1,5 @@
+export interface GlobalScriptStyleParsedEntry {
+    paths: string[];
+    entry: string;
+    lazy?: boolean;
+}
