+++ src/models/internals/index.ts
@@ -1,0 +1,12 @@
+export * from './angular-build-config-internal';
+export * from './app-project-config-internal';
+export * from './build-context-instance-options';
+export * from './build-context-static-options';
+export * from './build-options-internal';
+export * from './global-script-style-parsed-entry';
+export * from './lib-bundle-options-internal';
+export * from './lib-project-config-internal';
+export * from './package-entrypoints';
+export * from './polyfill-dll-parsed-entry';
+export * from './project-config-internal';
+export * from './ts-transpilation-options-internal';
