+++ src/models/internals/polyfill-dll-parsed-entry.ts
@@ -1,0 +1,5 @@
+export interface PolyfillDllParsedEntry {
+    tsEntries: string[];
+    scriptEntries: string[];
+    styleEntries: string[];
+}
