+++ src/models/project-config-compat.ts
